+++ src/Phan/Bootstrap.php
@@ -287,8 +287,9 @@ function phan_error_handler(int $errno, string $errstr, string $errfile, int $er
         // Don't execute the PHP internal error handler.
         return true;
     }
-    if ($errno === E_DEPRECATED && preg_match('/^Method ReflectionParameter::getClass/', $errstr)) {
+    if ($errno === E_DEPRECATED && preg_match('/^(Constant |Method ReflectionParameter::getClass)/', $errstr)) {
         // Suppress deprecation notices running `vendor/bin/paratest` in php 8
+        // Constants such as ENCHANT can be deprecated when calling constant()
         return true;
     }
     if ($errno === E_DEPRECATED && preg_match('/ast\\\\parse_.*Version.*is deprecated/i', $errstr)) {