+++ tests/php80_files/src/010_noncapturing_catch.php
@@ -1,0 +1,18 @@
+<?php
+try {
+    throw new RuntimeException();
+} catch (RuntimeException) {
+    echo "Caught\n";
+} catch (UnknownException) {
+    echo "Impossible\n";
+}
+function myFunction() {
+    try {
+        throw new RuntimeException();
+    } catch (RuntimeException) {
+        echo "Caught\n";
+    } catch (UnknownException) {
+        echo "Impossible\n";
+        throw new InvalidArgumentException();
+    }
+}